+++ src/FloppyBot.Commands.Core/Executor/CommandExecutor.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Immutable;
 using FloppyBot.Chat.Entities;
 using FloppyBot.Commands.Core.Entities;
+using FloppyBot.Commands.Core.Spawner;
 using FloppyBot.Commands.Parser.Entities;
 using Microsoft.Extensions.Logging;
