+++ src/FloppyBot.Commands.Core/Spawner/ICommandSpawner.cs
@@ -2,7 +2,7 @@
 using FloppyBot.Commands.Core.Entities;
 using FloppyBot.Commands.Parser.Entities;
 
-namespace FloppyBot.Commands.Core.Executor;
+namespace FloppyBot.Commands.Core.Spawner;
 
 public interface ICommandSpawner
 {
